// frontend_restaurants/web/src/lib/types.ts
export type StatisticChartType = {
	type: string;
	year: number;
	month?: number;
	week?: number;
	day?: number;
	timeOfDay: string;
};

export type TimeRangeSelectionProps = {
	timeRange: StatisticChartType;
	setTimeRange: React.Dispatch<React.SetStateAction<StatisticChartType>>;
};